+++ src/utilities/index.ts
@@ -8,3 +8,5 @@ export {getTextContainerHeight} from './get-text-container-height';
 export {RightAngleTriangle} from './RightAngleTriangle';
 export {degreesToRadians} from './degrees-to-radians';
 export {getBarXAxisDetails} from './get-bar-xaxis-details';
+export {getLongestLabelDetails} from './get-longest-label-details';
+export {getMaxDiagonalDetails} from './get-max-diagonal-details';